More refactor and cleanup
